+++ packages/user-onboarding/projects/user-onboarding-lib/src/lib/tour-service.service.ts
@@ -26,10 +26,10 @@ export class TourServiceService {
   constructor(
     private readonly tourStoreService: TourStoreServiceService,
     private readonly router: Router
-  ) {}
+  ) { }
 
   private addRemovedSteps(removedSteps): void {
-    for (const step of removedSteps) {
+    for (let i = 0; i < removedSteps.length; i++) {
       this.tour.steps.splice(0, 0, this.tour.steps.pop());
     }
   }
