+++ packages/user-onboarding/projects/user-onboarding-lib/src/test.ts
@@ -1,11 +1,11 @@
 // This file is required by karma.conf.js and loads recursively all the .spec and framework files
 
-import 'zone.js/dist/zone-testing';
+import 'zone.js/dist/zone-testing'; //NOSONAR
 import { getTestBed } from '@angular/core/testing';
 import {
   BrowserDynamicTestingModule,
   platformBrowserDynamicTesting
-} from '@angular/platform-browser-dynamic/testing';
+} from '@angular/platform-browser-dynamic/testing'; //NOSONAR
 
 declare const require: {
   context(path: string, deep?: boolean, filter?: RegExp): {
