+++ sandbox/auth-multitenant-example/src/providers/casbin-authorization-action.provider.ts
@@ -1,6 +1,6 @@
-import {Provider, inject, Getter} from '@loopback/core';
-import {HttpErrors, Request} from '@loopback/rest';
-import {IAuthUserWithPermissions, ILogger, LOGGER} from '@sourceloop/core';
+import { Provider, inject, Getter } from '@loopback/core';
+import { HttpErrors, Request } from '@loopback/rest';
+import { IAuthUserWithPermissions, ILogger, LOGGER } from '@sourceloop/core';
 import {
   CasbinAuthorizeFn,
   AuthorizationBindings,
@@ -24,7 +24,7 @@ export class CasbinAuthorizationProvider
     @inject(AuthorizationBindings.PATHS_TO_ALLOW_ALWAYS)
     private readonly allowAlwaysPath: string[],
     @inject(LOGGER.LOGGER_INJECT) public logger: ILogger,
-  ) {}
+  ) { }
 
   value(): CasbinAuthorizeFn {
     return (response, resource, request) =>
@@ -65,18 +65,13 @@ export class CasbinAuthorizationProvider
       const enforcer = await this.getEnforcer(metadata, casbinConfig, subject);
       // If casbin config policy format is being used, create enforcer
       const resourceIds = resource.split(',');
-      for (const resourceId of resourceIds) {
-        for (const permission of desiredPermissions) {
-          authDecision = await enforcer.enforce(
-            subject,
-            permission,
-            resourceId,
-          );
-          if (authDecision) {
-            break;
-          }
-        }
-      }
+      authDecision = await this.makeDecision(
+        resourceIds,
+        desiredPermissions,
+        enforcer,
+        subject,
+        authDecision
+      );
     } catch (err) {
       this.logger.error(`Error while doing casbin action. Error :: ${err}`);
       throw new HttpErrors.Forbidden(AuthorizeErrorKeys.NotAllowedAccess);
@@ -92,6 +87,28 @@ export class CasbinAuthorizationProvider
     return `u${id}`;
   }
 
+  async makeDecision(
+    resourceIds: string[],
+    desiredPermissions: string[],
+    enforcer: casbin.Enforcer,
+    subject: string,
+    authDecision?: boolean
+  ) {
+    let decision = authDecision ?? false;
+    for (const resourceId of resourceIds) {
+      for (const permission of desiredPermissions) {
+        decision = await enforcer.enforce(
+          subject,
+          permission,
+          resourceId,
+        );
+        if (decision) {
+          break;
+        }
+      }
+    }
+    return decision;
+  }
   createCasbinPolicy(
     resPermObj: ResourcePermissionObject[],
     subject: string,
