+++ sandbox/notification-socket-example/public/index.html
@@ -15,7 +15,8 @@
       text-align: center;
     }
 
-    a, a:visited {
+    a,
+    a:visited {
       color: #3f5dff;
     }
 
@@ -23,7 +24,9 @@
       margin-left: 10px;
     }
 
-    a:hover, a:focus, a:active {
+    a:hover,
+    a:focus,
+    a:active {
       color: #001956;
     }
 
@@ -58,11 +61,14 @@
         color: white;
       }
 
-      a, a:visited {
+      a,
+      a:visited {
         color: #4990e2;
       }
 
-      a:hover, a:focus, a:active {
+      a:hover,
+      a:focus,
+      a:active {
         color: #2b78ff;
       }
     }
@@ -80,7 +86,8 @@
 
   <footer class="power">
     <a href="https://loopback.io" target="_blank">
-      <img src="https://loopback.io/images/branding/powered-by-loopback/blue/powered-by-loopback-sm.png" />
+      <img src="https://loopback.io/images/branding/powered-by-loopback/blue/powered-by-loopback-sm.png"
+        alt="loopback logo" />
     </a>
   </footer>
 </body>
