+++ sandbox/notification-socket-example/src/controllers/ping.controller.ts
@@ -1,6 +1,6 @@
-import {Request, RestBindings, get, ResponseObject} from '@loopback/rest';
-import {inject} from '@loopback/core';
-import {authorize} from 'loopback4-authorization';
+import { Request, RestBindings, get, ResponseObject } from '@loopback/rest';
+import { inject } from '@loopback/core';
+import { authorize } from 'loopback4-authorization';
 
 /**
  * OpenAPI response for ping()
@@ -13,13 +13,13 @@ const PING_RESPONSE: ResponseObject = {
         type: 'object',
         title: 'PingResponse',
         properties: {
-          greeting: {type: 'string'},
-          date: {type: 'string'},
-          url: {type: 'string'},
+          greeting: { type: 'string' },
+          date: { type: 'string' },
+          url: { type: 'string' },
           headers: {
             type: 'object',
             properties: {
-              'Content-Type': {type: 'string'},
+              'Content-Type': { type: 'string' },
             },
             additionalProperties: true,
           },
@@ -33,10 +33,10 @@ const PING_RESPONSE: ResponseObject = {
  * A simple controller to bounce back http requests
  */
 export class PingController {
-  constructor(@inject(RestBindings.Http.REQUEST) private req: Request) {}
+  constructor(@inject(RestBindings.Http.REQUEST) private readonly req: Request) { }
 
   // Map to `GET /ping`
-  @authorize({permissions: ['*']})
+  @authorize({ permissions: ['*'] })
   @get('/ping', {
     responses: {
       '200': PING_RESPONSE,
