+++ sandbox/notification-socket-example/src/openapi-spec.ts
@@ -1,5 +1,5 @@
-import {ApplicationConfig} from '@loopback/core';
-import {NotificationSocketExampleApplication} from './application';
+import { ApplicationConfig } from '@loopback/core';
+import { NotificationSocketExampleApplication } from './application';
 
 /**
  * Export the OpenAPI spec from the application
@@ -11,7 +11,6 @@ async function exportOpenApiSpec(): Promise<void> {
       host: process.env.HOST ?? 'localhost',
     },
   };
-  // const outFile = process.argv[2] ?? '';
   const app = new NotificationSocketExampleApplication(config);
   await app.boot();
   // await app.exportOpenApiSpec(outFile);
