+++ services/authentication-service/package.json
@@ -58,12 +58,14 @@
     "@loopback/repository": "^3.6.0",
     "@loopback/rest": "^9.3.0",
     "@loopback/rest-explorer": "^3.3.0",
+    "@loopback/service-proxy": "^3.2.2",
     "@sourceloop/core": "^2.0.0",
     "@types/i18n": "^0.8.6",
     "base-64": "^1.0.0",
     "bcrypt": "^5.0.0",
     "dotenv": "^8.2.0",
     "dotenv-extended": "^2.8.0",
+    "https-proxy-agent": "^5.0.0",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
     "loopback4-authentication": "^4.6.0",
