+++ services/authentication-service/src/modules/auth/login.controller.ts
@@ -4,13 +4,11 @@ import {
   get,
   getModelSchemaRef,
   HttpErrors,
-  oas,
   param,
   patch,
   post,
   Request,
   requestBody,
-  Response,
   RestBindings,
 } from '@loopback/rest';
 import {
@@ -37,16 +35,10 @@ import {
 } from 'loopback4-authentication';
 import {authorize, AuthorizeErrorKeys} from 'loopback4-authorization';
 import moment from 'moment-timezone';
-import {URLSearchParams} from 'url';
 
 import {AuthServiceBindings} from '../../keys';
 import {AuthClient, RefreshToken, User} from '../../models';
-import {
-  AuthCodeBindings,
-  CodeReaderFn,
-  CodeWriterFn,
-  JwtPayloadFn,
-} from '../../providers';
+import {AuthCodeBindings, CodeReaderFn, JwtPayloadFn} from '../../providers';
 import {
   AuthClientRepository,
   RefreshTokenRepository,
@@ -60,17 +52,8 @@ import {
 import {TenantConfigRepository} from '../../repositories/tenant-config.repository';
 import {AuthRefreshTokenRequest, AuthTokenRequest, LoginRequest} from './';
 import {AuthUser, DeviceInfo} from './models/auth-user.model';
-import {ClientAuthRequest} from './models/client-auth-request.dto';
 import {ResetPassword} from './models/reset-password.dto';
 import {TokenResponse} from './models/token-response.dto';
-
-const queryGen = (from: 'body' | 'query') => {
-  return (req: Request) => {
-    return {
-      state: `client_id=${req[from].client_id}`,
-    };
-  };
-};
 
 const userAgentKey = 'user-agent';
 
@@ -138,7 +121,7 @@ export class LoginController {
     });
     if (!this.client || !this.user) {
       throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    } else if (!req.client_secret) {
+    } else if (!req.clientSecret) {
       throw new HttpErrors.BadRequest(AuthErrorKeys.ClientSecretMissing);
     } else if (userStatus?.status === UserStatus.REGISTERED) {
       throw new HttpErrors.BadRequest('User not active yet');
@@ -147,12 +130,12 @@ export class LoginController {
     }
     try {
       const codePayload: ClientAuthCode<User, typeof User.prototype.id> = {
-        clientId: req.client_id,
+        clientId: req.clientId,
         userId: this.user.id,
       };
       const token = jwt.sign(codePayload, this.client.secret, {
         expiresIn: this.client.authCodeExpiration,
-        audience: req.client_id,
+        audience: req.clientId,
         issuer: process.env.JWT_ISSUER,
         algorithm: 'HS256',
       });
@@ -202,7 +185,7 @@ export class LoginController {
       this.user.authClientIds.length === 0
     ) {
       throw new HttpErrors.UnprocessableEntity(AuthErrorKeys.ClientUserMissing);
-    } else if (!req.client_secret) {
+    } else if (!req.clientSecret) {
       throw new HttpErrors.BadRequest(AuthErrorKeys.ClientSecretMissing);
     } else if (userStatus?.status === UserStatus.REGISTERED) {
       throw new HttpErrors.BadRequest('Your Sign-Up request is in process');
@@ -344,598 +327,6 @@ export class LoginController {
         userAgent: this.req.headers[userAgentKey],
       },
     );
-  }
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.GOOGLE_OAUTH2,
-    {
-      accessType: 'offline',
-      scope: ['profile', 'email'],
-      authorizationURL: process.env.GOOGLE_AUTH_URL,
-      callbackURL: process.env.GOOGLE_AUTH_CALLBACK_URL,
-      clientID: process.env.GOOGLE_AUTH_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.GOOGLE_AUTH_TOKEN_URL,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @oas.deprecated()
-  @get('/auth/google', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description:
-          'Google Token Response (Deprecated: Possible security issue if secret is passed via query params, please use the post endpoint)',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  /**
-   * @deprecated This method should not be used, possible security issue if secret is passed via query params, please use the post endpoint
-   */
-  async loginViaGoogle(
-    @param.query.string('client_id')
-    clientId?: string,
-    @param.query.string('client_secret')
-    clientSecret?: string,
-  ): Promise<void> {}
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.GOOGLE_OAUTH2,
-    {
-      accessType: 'offline',
-      scope: ['profile', 'email'],
-      authorizationURL: process.env.GOOGLE_AUTH_URL,
-      callbackURL: process.env.GOOGLE_AUTH_CALLBACK_URL,
-      clientID: process.env.GOOGLE_AUTH_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.GOOGLE_AUTH_TOKEN_URL,
-    },
-    queryGen('body'),
-  )
-  @authorize({permissions: ['*']})
-  @post('/auth/google', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'POST Call for Google based login',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async postLoginViaGoogle(
-    @requestBody({
-      content: {
-        'application/x-www-form-urlencoded': {
-          schema: getModelSchemaRef(ClientAuthRequest),
-        },
-      },
-    })
-    clientCreds?: ClientAuthRequest,
-  ): Promise<void> {}
-
-  @authenticate(
-    STRATEGY.GOOGLE_OAUTH2,
-    {
-      accessType: 'offline',
-      scope: ['profile', 'email'],
-      authorizationURL: process.env.GOOGLE_AUTH_URL,
-      callbackURL: process.env.GOOGLE_AUTH_CALLBACK_URL,
-      clientID: process.env.GOOGLE_AUTH_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.GOOGLE_AUTH_TOKEN_URL,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @get('/auth/google-auth-redirect', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'Google Redirect Token Response',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async googleCallback(
-    @param.query.string('code') code: string,
-    @param.query.string('state') state: string,
-    @inject(RestBindings.Http.RESPONSE) response: Response,
-    @inject(AuthCodeBindings.CODEWRITER_PROVIDER)
-    googleCodeWriter: CodeWriterFn,
-  ): Promise<void> {
-    const clientId = new URLSearchParams(state).get('client_id');
-    if (!clientId || !this.user) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    const client = await this.authClientRepository.findOne({
-      where: {
-        clientId,
-      },
-    });
-    if (!client || !client.redirectUrl) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    try {
-      const codePayload: ClientAuthCode<User, typeof User.prototype.id> = {
-        clientId,
-        user: this.user,
-      };
-      const token = await googleCodeWriter(
-        jwt.sign(codePayload, client.secret, {
-          expiresIn: client.authCodeExpiration,
-          audience: clientId,
-          issuer: process.env.JWT_ISSUER,
-          algorithm: 'HS256',
-        }),
-      );
-      response.redirect(`${client.redirectUrl}?code=${token}`);
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
-    }
-  }
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.INSTAGRAM_OAUTH2,
-    {
-      accessType: 'offline',
-      authorizationURL: process.env.INSTAGRAM_AUTH_URL,
-      callbackURL: process.env.INSTAGRAM_AUTH_CALLBACK_URL,
-      clientID: process.env.INSTAGRAM_AUTH_CLIENT_ID,
-      clientSecret: process.env.INSTAGRAM_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.INSTAGRAM_AUTH_TOKEN_URL,
-    },
-    queryGen('body'),
-  )
-  @authorize({permissions: ['*']})
-  @post('/auth/instagram', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'POST Call for Instagram based login',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async postLoginViaInstagram(
-    @requestBody({
-      content: {
-        'application/x-www-form-urlencoded': {
-          schema: getModelSchemaRef(ClientAuthRequest),
-        },
-      },
-    })
-    clientCreds?: ClientAuthRequest,
-  ): Promise<void> {}
-
-  @authenticate(
-    STRATEGY.INSTAGRAM_OAUTH2,
-    {
-      accessType: 'offline',
-      authorizationURL: process.env.INSTAGRAM_AUTH_URL,
-      callbackURL: process.env.INSTAGRAM_AUTH_CALLBACK_URL,
-      clientID: process.env.INSTAGRAM_AUTH_CLIENT_ID,
-      clientSecret: process.env.INSTAGRAM_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.INSTAGRAM_AUTH_TOKEN_URL,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @get('/auth/instagram-auth-redirect', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'Instagram Redirect Token Response',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async instagramCallback(
-    @param.query.string('code') code: string,
-    @param.query.string('state') state: string,
-    @inject(RestBindings.Http.RESPONSE) response: Response,
-    @inject(AuthCodeBindings.CODEWRITER_PROVIDER)
-    instgaramCodeWriter: CodeWriterFn,
-  ): Promise<void> {
-    const clientId = new URLSearchParams(state).get('client_id');
-    if (!clientId || !this.user) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    const client = await this.authClientRepository.findOne({
-      where: {
-        clientId,
-      },
-    });
-    if (!client || !client.redirectUrl) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    try {
-      const codePayload: ClientAuthCode<User, typeof User.prototype.id> = {
-        clientId,
-        user: this.user,
-      };
-      const token = await instgaramCodeWriter(
-        jwt.sign(codePayload, client.secret, {
-          expiresIn: client.authCodeExpiration,
-          audience: clientId,
-          issuer: process.env.JWT_ISSUER,
-          algorithm: 'HS256',
-        }),
-      );
-      response.redirect(`${client.redirectUrl}?code=${token}`);
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
-    }
-  }
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.APPLE_OAUTH2,
-    {
-      accessType: 'offline',
-      scope: ['name', 'email'],
-      callbackURL: process.env.APPLE_AUTH_CALLBACK_URL,
-      clientID: process.env.APPLE_AUTH_CLIENT_ID,
-      teamID: process.env.APPLE_AUTH_TEAM_ID,
-      keyID: process.env.APPLE_AUTH_KEY_ID,
-      privateKeyLocation: process.env.APPLE_AUTH_PRIVATE_KEY_LOCATION,
-    },
-    queryGen('body'),
-  )
-  @authorize({permissions: ['*']})
-  @post('/auth/oauth-apple', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'POST Call for Apple based login',
-        content: {},
-      },
-    },
-  })
-  postLoginViaApple(
-    @requestBody({
-      content: {
-        'application/x-www-form-urlencoded': {
-          schema: getModelSchemaRef(ClientAuthRequest),
-        },
-      },
-    })
-    clientCreds: ClientAuthRequest,
-  ): void {}
-
-  @authenticate(
-    STRATEGY.APPLE_OAUTH2,
-    {
-      accessType: 'offline',
-      scope: ['name', 'email'],
-      callbackURL: process.env.APPLE_AUTH_CALLBACK_URL,
-      clientID: process.env.APPLE_AUTH_CLIENT_ID,
-      teamID: process.env.APPLE_AUTH_TEAM_ID,
-      keyID: process.env.APPLE_AUTH_KEY_ID,
-      privateKeyLocation: process.env.APPLE_AUTH_PRIVATE_KEY_LOCATION,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @get('/auth/apple-oauth-redirect', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'Apple Redirect Token Response',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async appleCallback(
-    @param.query.string('code') code: string,
-    @param.query.string('state') state: string,
-    @inject(RestBindings.Http.RESPONSE) response: Response,
-    @inject(AuthCodeBindings.CODEWRITER_PROVIDER)
-    appleCodeWriter: CodeWriterFn,
-  ): Promise<void> {
-    const clientId = new URLSearchParams(state).get('client_id');
-
-    if (!clientId || !this.user) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    const client = await this.authClientRepository.findOne({
-      where: {
-        clientId,
-      },
-    });
-    if (!client || !client.redirectUrl) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    try {
-      const codePayload: ClientAuthCode<User, typeof User.prototype.id> = {
-        clientId,
-        user: this.user,
-      };
-      const token = await appleCodeWriter(
-        jwt.sign(codePayload, client.secret, {
-          expiresIn: client.authCodeExpiration,
-          audience: clientId,
-          issuer: process.env.JWT_ISSUER,
-          algorithm: 'HS256',
-        }),
-      );
-      const role = this.user.role;
-      response.redirect(
-        `${process.env.WEBAPP_URL ?? ''}${
-          client.redirectUrl
-        }?code=${token}&role=${role}`,
-      );
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
-    }
-  }
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.FACEBOOK_OAUTH2,
-    {
-      accessType: 'offline',
-      authorizationURL: process.env.FACEBOOK_AUTH_URL,
-      callbackURL: process.env.FACEBOOK_AUTH_CALLBACK_URL,
-      clientID: process.env.FACEBOOK_AUTH_CLIENT_ID,
-      clientSecret: process.env.FACEBOOK_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.FACEBOOK_AUTH_TOKEN_URL,
-    },
-    queryGen('body'),
-  )
-  @authorize({permissions: ['*']})
-  @post('/auth/facebook', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'POST Call for Facebook based login',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async postLoginViaFacebook(
-    @requestBody({
-      content: {
-        'application/x-www-form-urlencoded': {
-          schema: getModelSchemaRef(ClientAuthRequest),
-        },
-      },
-    })
-    clientCreds?: ClientAuthRequest,
-  ): Promise<void> {}
-
-  @authenticate(
-    STRATEGY.FACEBOOK_OAUTH2,
-    {
-      accessType: 'offline',
-      authorizationURL: process.env.FACEBOOK_AUTH_URL,
-      callbackURL: process.env.FACEBOOK_AUTH_CALLBACK_URL,
-      clientID: process.env.FACEBOOK_AUTH_CLIENT_ID,
-      clientSecret: process.env.FACEBOOK_AUTH_CLIENT_SECRET,
-      tokenURL: process.env.FACEBOOK_AUTH_TOKEN_URL,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @get('/auth/facebook-auth-redirect', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'Facebook Redirect Token Response',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async facebookCallback(
-    @param.query.string('code') code: string,
-    @param.query.string('state') state: string,
-    @inject(RestBindings.Http.RESPONSE) response: Response,
-    @inject(AuthCodeBindings.CODEWRITER_PROVIDER)
-    facebookCodeWriter: CodeWriterFn,
-  ): Promise<void> {
-    const clientId = new URLSearchParams(state).get('client_id');
-    if (!clientId || !this.user) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    const client = await this.authClientRepository.findOne({
-      where: {
-        clientId,
-      },
-    });
-    if (!client || !client.redirectUrl) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    try {
-      const codePayload: ClientAuthCode<User, typeof User.prototype.id> = {
-        clientId,
-        user: this.user,
-      };
-      const token = await facebookCodeWriter(
-        jwt.sign(codePayload, client.secret, {
-          expiresIn: client.authCodeExpiration,
-          audience: clientId,
-          issuer: process.env.JWT_ISSUER,
-          algorithm: 'HS256',
-        }),
-      );
-      response.redirect(`${client.redirectUrl}?code=${token}`);
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
-    }
-  }
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.KEYCLOAK,
-    {
-      host: process.env.KEYCLOAK_HOST,
-      realm: process.env.KEYCLOAK_REALM, //'Tenant1',
-      clientID: process.env.KEYCLOAK_CLIENT_ID, //'onboarding',
-      clientSecret: process.env.KEYCLOAK_CLIENT_SECRET, //'e607fd75-adc8-4af7-9f03-c9e79a4b8b72',
-      callbackURL: process.env.KEYCLOAK_CALLBACK_URL, //'http://localhost:3001/auth/keycloak-auth-redirect',
-      authorizationURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/auth`,
-      tokenURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/token`,
-      userInfoURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/userinfo`,
-    },
-    queryGen('body'),
-  )
-  @authorize({permissions: ['*']})
-  @post('/auth/keycloak', {
-    description: 'POST Call for keycloak based login',
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'Keycloak Token Response',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async postLoginViaKeycloak(
-    @requestBody({
-      content: {
-        'application/x-www-form-urlencoded': {
-          schema: getModelSchemaRef(ClientAuthRequest),
-        },
-      },
-    })
-    clientCreds?: ClientAuthRequest,
-  ): Promise<void> {}
-
-  @authenticateClient(STRATEGY.CLIENT_PASSWORD)
-  @authenticate(
-    STRATEGY.KEYCLOAK,
-    {
-      host: process.env.KEYCLOAK_HOST,
-      realm: process.env.KEYCLOAK_REALM, //'Tenant1',
-      clientID: process.env.KEYCLOAK_CLIENT_ID, //'onboarding',
-      clientSecret: process.env.KEYCLOAK_CLIENT_SECRET, //'e607fd75-adc8-4af7-9f03-c9e79a4b8b72',
-      callbackURL: process.env.KEYCLOAK_CALLBACK_URL, //'http://localhost:3001/auth/keycloak-auth-redirect',
-      authorizationURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/auth`,
-      tokenURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/token`,
-      userInfoURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/userinfo`,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @oas.deprecated()
-  @get('/auth/keycloak', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description:
-          'Keycloak Token Response (Deprecated: Possible security issue if secret is passed via query params, please use the post endpoint)',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async loginViaKeycloak(
-    @param.query.string('client_id')
-    clientId?: string,
-    @param.query.string('client_secret')
-    clientSecret?: string,
-  ): Promise<void> {}
-
-  @authenticate(
-    STRATEGY.KEYCLOAK,
-    {
-      host: process.env.KEYCLOAK_HOST,
-      realm: process.env.KEYCLOAK_REALM, //'Tenant1',
-      clientID: process.env.KEYCLOAK_CLIENT_ID, //'onboarding',
-      clientSecret: process.env.KEYCLOAK_CLIENT_SECRET, //'e607fd75-adc8-4af7-9f03-c9e79a4b8b72',
-      callbackURL: process.env.KEYCLOAK_CALLBACK_URL, //'http://localhost:3001/auth/keycloak-auth-redirect',
-      authorizationURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/auth`,
-      tokenURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/token`,
-      userInfoURL: `${process.env.KEYCLOAK_HOST}/auth/realms/${process.env.KEYCLOAK_REALM}/protocol/openid-connect/userinfo`,
-    },
-    queryGen('query'),
-  )
-  @authorize({permissions: ['*']})
-  @get('/auth/keycloak-auth-redirect', {
-    responses: {
-      [STATUS_CODE.OK]: {
-        description: 'Keycloak Redirect Token Response',
-        content: {
-          [CONTENT_TYPE.JSON]: {
-            schema: {'x-ts-type': TokenResponse},
-          },
-        },
-      },
-    },
-  })
-  async keycloakCallback(
-    @param.query.string('code') code: string,
-    @param.query.string('state') state: string,
-    @inject(RestBindings.Http.RESPONSE) response: Response,
-    @inject(AuthCodeBindings.CODEWRITER_PROVIDER)
-    keycloackCodeWriter: CodeWriterFn,
-  ): Promise<void> {
-    const clientId = new URLSearchParams(state).get('client_id');
-    if (!clientId || !this.user) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    const client = await this.authClientRepository.findOne({
-      where: {
-        clientId,
-      },
-    });
-    if (!client || !client.redirectUrl) {
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.ClientInvalid);
-    }
-    try {
-      const codePayload: ClientAuthCode<User, typeof User.prototype.id> = {
-        clientId,
-        user: this.user,
-      };
-      const token = await keycloackCodeWriter(
-        jwt.sign(codePayload, client.secret, {
-          expiresIn: client.authCodeExpiration,
-          audience: clientId,
-          issuer: process.env.JWT_ISSUER,
-          algorithm: 'HS256',
-        }),
-      );
-      response.redirect(`${client.redirectUrl}?code=${token}`);
-    } catch (error) {
-      this.logger.error(error);
-      throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
-    }
   }
 
   @authenticate(STRATEGY.BEARER, {passReqToCallback: true})
