+++ services/scheduler-service/package.json
@@ -61,6 +61,7 @@
     "@loopback/repository": "^3.6.0",
     "@loopback/rest": "^9.3.0",
     "@loopback/rest-explorer": "^3.3.0",
+    "@loopback/service-proxy": "^3.2.2",
     "@sourceloop/core": "^2.0.0",
     "dotenv": "^8.2.0",
     "dotenv-extended": "^2.9.0",
@@ -68,6 +69,7 @@
     "loopback4-authorization": "^3.2.0",
     "loopback4-soft-delete": "^3.2.1",
     "prom-client": "^13.1.0",
+    "sinon": "^11.1.1",
     "tslib": "^2.0.0"
   },
   "devDependencies": {
